refactor(core): 迁移 gateway 管理路由到 admin 域
