--- noj-core/src/domains/gateway/routes/index.ts
@@ -1,6 +1,0 @@
-import { Hono } from "hono";
-import adminLlm from "./admin-llm.ts";
-
-/** gateway 域管理路由，挂载到 `/api/v1/admin`。 */
-export const gatewayAdminRouter = new Hono();
-gatewayAdminRouter.route("/", adminLlm);
