+++ noj-core/tests/routes/admin-llm.test.ts
@@ -9,7 +9,7 @@ import {
   ServiceUnavailableError,
 } from "../../src/shared/base/errors.ts";
 import { LlmGatewayError } from "../../src/domains/gateway/services/llm.ts";
-import { mapLlmError } from "../../src/domains/gateway/routes/admin-llm.ts";
+import { mapLlmError } from "../../src/domains/admin/routes/gateway.ts";
 
 Deno.test("admin-llm: gateway 404 映射为 NotFoundError", () => {
   const error = assertThrows(() =>
